feat: Concurrent, Isolated Processes, exercise 3
Transfer Socket Ownership

When a socket is created, it remembers the process that created it. That
process is referred to as the controlling process. And if that
controlling process dies, it takes care of closing the socket it
created. If a spawned process dies, it doesn't automatically close the
client_socket because the spawned process is not the controlling
process. This makes the spawned process also the controlling process so
the socket connection will be closed.

iex -S mix
spawn(Servy.HttpServer, :start, [4000])

then in separate terminal from the root of the project directory
❯ curl http://localhost:4000/kaboom
curl: (52) Empty reply from server